+++ src/pyjams/color/__init__.py
@@ -29,8 +29,9 @@ from .pyjams_palettes import pyjams_cmaps
 from .sron2012_palettes import sron2012_colors, sron2012_functions
 from .sron_palettes import sron_colors, sron_colormaps, sron_functions
 from .ufz_palettes import ufz_colors
-# get, show, print color palettes
-from .color import get_color, get_cmap, print_palettes, show_palettes
+# get, show, print colors and color palettes
+from .color import get_color, print_colors
+from .color import get_cmap, print_palettes, show_palettes
 
 
 __all__ = ['brewer_sequential', 'brewer_diverging', 'brewer_qualitative',
@@ -41,5 +42,6 @@ __all__ = ['brewer_sequential', 'brewer_diverging', 'brewer_qualitative',
            'sron2012_colors', 'sron2012_functions',
            'sron_colors', 'sron_colormaps', 'sron_functions',
            'ufz_colors',
-           'get_color', 'get_cmap', 'print_palettes', 'show_palettes',
+           'get_color', 'print_colors',
+           'get_cmap', 'print_palettes', 'show_palettes',
            ]